Add capability to POST a file (upload)
Enhance request processor to take a file resource in the data passed
into the request and automatically convert it to a string as well as add
the required headers and boundary necessary to upload the file.  Now,
passing an array with a key of a filename and the value as a resource
will upload the file with that filename.

An example to upload a file:
    $client = new Service_SmartFile_BasicClient(API_KEY, API_PWD);
    $rh = fopen("/etc/motd", "rb");
    $client->post("/path/data/", array("motd" => $rh));
    fclose($rh);

In addtion, change means to pass extra headers to request from a string
to an array.  The library now takes care of converting it into a string.